+++ src/lib/quests/index.ts
@@ -1,3 +1,4 @@
 export * from "./types";
 export * from "./transitions";
 export * from "./matchSkill";
+export * from "./evaluateQuest";